refactor: split A330 HLE services by responsibility
