+++ crates/dingooemu-core/src/a330/runtime.rs
@@ -19,6 +19,8 @@ use crate::package::PackageImage;
 use std::collections::{BTreeMap, BTreeSet, VecDeque};
 use std::path::PathBuf;
 
+mod sdk_hle;
+
 const INSTRUCTIONS_PER_SLICE: u64 = 1_000_000;
 const APP_PATH_ADDRESS: u32 = STACK_BASE + 0x200;
 const LOCALE_ADDRESS: u32 = STACK_BASE + 0x600;
@@ -765,853 +767,6 @@ struct RuntimeBus<'a> {
     console_output: &'a mut Vec<u8>,
 }
 
-impl RuntimeBus<'_> {
-    fn dispatch(&mut self, cpu: &mut Cpu, immediate: u32) -> Result<()> {
-        if immediate == 0x0012_3456 {
-            return self.dispatch_semihosting(cpu);
-        }
-        let (symbol_name, symbol_address) = if immediate & 0x0080_0000 != 0 {
-            let index = (immediate & 0x007f_ffff) as usize;
-            let name = self
-                .dynamic_imports
-                .get(index)
-                .ok_or_else(|| SimulatorError::CpuError {
-                    pc: cpu.r[15].wrapping_sub(4),
-                    message: format!("dynamic ARM SVC index {index} is invalid"),
-                })?;
-            (name.clone(), DYNAMIC_THUNK_BASE + index as u32 * 8)
-        } else {
-            let symbol =
-                self.imports
-                    .get(immediate as usize)
-                    .ok_or_else(|| SimulatorError::CpuError {
-                        pc: cpu.r[15].wrapping_sub(4),
-                        message: format!("ARM SVC index {immediate} is outside the import table"),
-                    })?;
-            (symbol.name.clone(), symbol.address)
-        };
-        let name = symbol_name.as_str();
-        log::trace!(
-            "ARM HLE {name}(r0={:#010x}, r1={:#010x}, r2={:#010x}, r3={:#010x})",
-            cpu.r[0],
-            cpu.r[1],
-            cpu.r[2],
-            cpu.r[3]
-        );
-        match name {
-            "lcd_get_frame" | "_lcd_get_frame" | "LCDGetFB" => cpu.r[0] = *self.active_framebuffer,
-            "LCDGetWidth" | "get_lcd_width" => cpu.r[0] = SCREEN_WIDTH,
-            "LCDGetHeight" | "get_lcd_height" => cpu.r[0] = SCREEN_HEIGHT,
-            "lcd_set_frame" | "_lcd_set_frame" | "LCDFlushFB" | "LCDFlushFBZoom" => {
-                *self.frame_address = Some(*self.active_framebuffer);
-                cpu.r[0] = 0;
-            }
-            "malloc" | "OSMalloc" | "jmalloc" => cpu.r[0] = self.allocate(cpu.r[0]),
-            "calloc" => {
-                cpu.r[0] = match cpu.r[0].checked_mul(cpu.r[1]) {
-                    Some(size) => self.allocate_zeroed(size)?,
-                    None => 0,
-                };
-            }
-            "realloc" => cpu.r[0] = self.reallocate(cpu.r[0], cpu.r[1])?,
-            "free" | "OSFree" | "jfree" => {
-                self.deallocate(cpu.r[0]);
-                cpu.r[0] = 0;
-            }
-            "memset" => {
-                let data = vec![cpu.r[1] as u8; cpu.r[2] as usize];
-                self.memory.write_bytes(cpu.r[0], &data)?;
-            }
-            "memcpy" | "memmove" => {
-                let data = self
-                    .memory
-                    .read_bytes(cpu.r[1], cpu.r[2] as usize)?
-                    .to_vec();
-                self.memory.write_bytes(cpu.r[0], &data)?;
-            }
-            "fopen" | "fsys_fopen" => {
-                let name = self.read_c_string(cpu.r[0], 1024)?;
-                let mode = self.read_c_string(cpu.r[1], 16)?;
-                cpu.r[0] = self.open_file(&name, &mode);
-            }
-            "fsys_fopenW" => {
-                let name = self.read_wide_string(cpu.r[0], 1024)?;
-                let mode = self.read_c_string(cpu.r[1], 16)?;
-                cpu.r[0] = self.open_file(&name, &mode);
-            }
-            "fclose" | "fsys_fclose" | "fsys_fcloseW" => {
-                cpu.r[0] = self.close_file(cpu.r[0]);
-            }
-            "fread" | "fsys_fread" => {
-                cpu.r[0] = self.read_file(cpu.r[0], cpu.r[1], cpu.r[2], cpu.r[3])?;
-            }
-            "fwrite" | "fsys_fwrite" => {
-                cpu.r[0] = self.write_file(cpu.r[0], cpu.r[1], cpu.r[2], cpu.r[3])?;
-            }
-            "fseek" | "fsys_fseek" => {
-                cpu.r[0] = self.seek_file(cpu.r[0], cpu.r[1] as i32, cpu.r[2]);
-            }
-            "ftell" | "fsys_ftell" => {
-                cpu.r[0] = self
-                    .files
-                    .get(&cpu.r[0])
-                    .map_or(u32::MAX, |file| file.position as u32);
-            }
-            "feof" | "fsys_feof" => {
-                cpu.r[0] = self
-                    .files
-                    .get(&cpu.r[0])
-                    .map_or(1, |file| u32::from(file.position >= file.data.len()));
-            }
-            "ferror" | "fsys_ferror" => cpu.r[0] = u32::from(!self.files.contains_key(&cpu.r[0])),
-            "fsys_findfirst" => {
-                let pattern = self.read_c_string(cpu.r[0], 1024)?;
-                cpu.r[0] = self.begin_file_search(&pattern, cpu.r[1], cpu.r[2])?;
-            }
-            "fsys_findnext" => cpu.r[0] = self.next_file_search(cpu.r[0])?,
-            "fsys_findclose" => {
-                self.file_searches.remove(&cpu.r[0]);
-                cpu.r[0] = 0;
-            }
-            "dl_res_open" => cpu.r[0] = self.open_resource([cpu.r[2], cpu.r[1], cpu.r[0]]),
-            "dl_res_get_size" => {
-                cpu.r[0] = self
-                    .files
-                    .get(&cpu.r[0])
-                    .map_or(0, |file| file.data.len() as u32)
-            }
-            "dl_res_get_data" => {
-                cpu.r[0] = self.read_resource(cpu.r[0], cpu.r[1], cpu.r[2], cpu.r[3])?;
-            }
-            "dl_res_close" => {
-                let handle = cpu.r[0];
-                if self.files.remove(&handle).is_some() {
-                    self.deallocate(handle);
-                }
-                cpu.r[0] = 0;
-            }
-            "_kbd_get_status" | "kbd_get_status" | "_rmt_get_status" | "rmt_get_status" => {
-                let address = cpu.r[0];
-                let (pressed, released, status) = self.input.take_status();
-                let pressed = map_a330_input(self.profile, pressed);
-                let released = map_a330_input(self.profile, released);
-                let status = map_a330_input(self.profile, status);
-                self.memory.write32(address, pressed)?;
-                self.memory.write32(address.wrapping_add(4), released)?;
-                self.memory.write32(address.wrapping_add(8), status)?;
-                cpu.r[0] = 0;
-            }
-            "_kbd_get_key" | "kbd_get_key" | "_rmt_get_key" | "rmt_get_key" | "sys_get_key"
-            | "KBDGetSKey" | "KBDGetSKeyStatus" | "RMTGetSKey" => {
-                cpu.r[0] = map_a330_input(self.profile, self.input.buttons());
-            }
-            "_sys_judge_event" | "sys_judge_event" => {
-                cpu.r[0] = u32::from(self.input.take_pending_event());
-            }
-            "printf" | "fprintf" => cpu.r[0] = 0,
-            "stricmp" | "strcasecmp" => {
-                let left = self.read_c_string(cpu.r[0], 4096)?;
-                let right = self.read_c_string(cpu.r[1], 4096)?;
-                cpu.r[0] = compare_ascii_case_insensitive(&left, &right) as u32;
-            }
-            "_waveout_open" | "waveout_open" => {
-                let address = cpu.r[0];
-                let config = AudioConfig::new(
-                    self.memory.read32(address)?,
-                    self.memory.read16(address + 4)?,
-                    self.memory.read8(address + 6)?,
-                    self.memory.read8(address + 7)?,
-                );
-                cpu.r[0] = u32::from(config.is_some_and(|config| self.audio.open(config)));
-            }
-            "waveout_write" => {
-                let buffer = cpu.r[1];
-                let count = cpu.r[2];
-                if count == 0 || count > 4 * 1024 * 1024 {
-                    cpu.r[0] = 0;
-                } else if !self.audio.can_write() && self.profile == ArmProfile::Retail {
-                    cpu.r[15] = cpu.r[15].wrapping_sub(4);
-                    self.yield_requested = true;
-                } else {
-                    let data = self.memory.read_bytes(buffer, count as usize)?;
-                    cpu.r[0] = u32::from(self.audio.write(data));
-                }
-            }
-            "waveout_try_write" => {
-                let count = cpu.r[2];
-                cpu.r[0] = if count == 0 || count > 4 * 1024 * 1024 || !self.audio.can_write() {
-                    0
-                } else {
-                    let data = self.memory.read_bytes(cpu.r[1], count as usize)?;
-                    u32::from(self.audio.write(data))
-                };
-            }
-            "waveout_can_write" | "waveout_can_write_nonblocking" | "pcm_can_write" => {
-                cpu.r[0] = u32::from(self.audio.can_write())
-            }
-            "waveout_close" | "waveout_close_at_once" => cpu.r[0] = u32::from(self.audio.close()),
-            "_waveout_set_volume" | "waveout_set_volume" => {
-                cpu.r[0] = u32::from(self.audio.set_volume(cpu.r[0]))
-            }
-            "HP_Mute_sw" | "waveout_mute" => {
-                cpu.r[0] = u32::from(self.audio.set_muted(cpu.r[0] != 0))
-            }
-            "vxGoHome" | "abort" | "av_end_thread" | "av_queue_abort" => self.stop_requested = true,
-            "OSTaskCreate" => {
-                if cpu.r[0] != 0 && cpu.r[2] != 0 {
-                    let mut task = Cpu::new(cpu.r[0], cpu.r[2], EXIT_ADDRESS);
-                    task.set_unknown_instruction_policy(cpu.unknown_instruction_policy());
-                    task.r[0] = cpu.r[1];
-                    task.start();
-                    self.tasks.push_back((task, cpu.r[3] & 0xff));
-                }
-                cpu.r[0] = 0;
-            }
-            "OSTaskQuery" => {
-                let priority = cpu.r[0] & 0xff;
-                cpu.r[0] = if priority == self.current_priority
-                    || self.tasks.iter().any(|(_, value)| *value == priority)
-                {
-                    0
-                } else {
-                    41
-                };
-            }
-            "OSTaskDel" => {
-                let priority = cpu.r[0] & 0xff;
-                if priority == self.current_priority {
-                    self.finish_current = true;
-                    cpu.r[0] = 0;
-                } else if let Some(index) =
-                    self.tasks.iter().position(|(_, value)| *value == priority)
-                {
-                    self.tasks.remove(index);
-                    cpu.r[0] = 0;
-                } else {
-                    cpu.r[0] = 41;
-                }
-            }
-            "OSSemCreate" => {
-                let initial = cpu.r[0];
-                let handle = self.allocate(16);
-                if handle != 0 {
-                    self.semaphores.insert(handle, initial);
-                }
-                cpu.r[0] = handle;
-            }
-            "OSSemPend" => {
-                let handle = cpu.r[0];
-                match self.semaphores.get_mut(&handle) {
-                    Some(count) if *count > 0 => {
-                        *count -= 1;
-                        if cpu.r[2] != 0 {
-                            self.memory.write8(cpu.r[2], 0)?;
-                        }
-                        cpu.r[0] = 0;
-                    }
-                    Some(_) if self.profile == ArmProfile::Retail => {
-                        cpu.r[15] = cpu.r[15].wrapping_sub(4);
-                        self.yield_requested = true;
-                    }
-                    Some(_) => {
-                        self.yield_requested = true;
-                        cpu.r[0] = 0;
-                    }
-                    None => {
-                        if cpu.r[2] != 0 {
-                            self.memory.write8(cpu.r[2], 4)?;
-                        }
-                        cpu.r[0] = 0;
-                    }
-                }
-            }
-            "OSSemPost" => {
-                cpu.r[0] = match self.semaphores.get_mut(&cpu.r[0]) {
-                    Some(count) => {
-                        *count = count.saturating_add(1);
-                        0
-                    }
-                    None => 41,
-                };
-            }
-            "OSSemDel" => {
-                cpu.r[0] = if self.semaphores.remove(&cpu.r[0]).is_some() {
-                    0
-                } else {
-                    41
-                };
-            }
-            "OSTimeDly" | "delay" | "delay_ms" | "OSTimeDlyHMSM" => {
-                cpu.r[0] = 0;
-                self.yield_requested = true;
-            }
-            "OSTimeGet" => cpu.r[0] = (cpu.instruction_count / 150_000) as u32,
-            "GetTickCount" => cpu.r[0] = (cpu.instruction_count / 15_000) as u32,
-            "OSTimerGetTickTimeus" => cpu.r[0] = (cpu.instruction_count / 15) as u32,
-            "TaskMediaFunStop" => cpu.r[0] = 0,
-            "get_current_language" => cpu.r[0] = 0,
-            "GetDLHandle" | "get_dl_handle" => cpu.r[0] = STACK_BASE + 0x100,
-            "__to_locale_ansi" | "_to_locale_ansi" => cpu.r[0] = LOCALE_ADDRESS,
-            "dl_get_proc" => cpu.r[0] = self.dynamic_import(cpu.r[1])?,
-            "cmGetSysModel" => {
-                cpu.r[0] = u32::from(!self.write_guest_string(cpu.r[0], cpu.r[1], "CC1800")?)
-            }
-            "cmGetSysVersion" => {
-                cpu.r[0] = u32::from(!self.write_guest_string(cpu.r[0], cpu.r[1], "1.0")?)
-            }
-            "LCDIsDoubleFBEnabled" => cpu.r[0] = 1,
-            "LCDGetFBFormat" => cpu.r[0] = 0,
-            "LCDSetFBBit" => {
-                if matches!(cpu.r[0], 16 | 32) {
-                    *self.framebuffer_bits = cpu.r[0] | FRAMEBUFFER_BITS_EXPLICIT;
-                }
-                cpu.r[0] = 0;
-            }
-            "BMF_SetLcdFramePtr" => {
-                if self.memory.read_bytes(cpu.r[0], FRAMEBUFFER_SIZE).is_ok() {
-                    *self.active_framebuffer = cpu.r[0];
-                }
-                cpu.r[0] = 0;
-            }
-            "SysLcdClear" => {
-                let clear = vec![0; FRAMEBUFFER_SIZE];
-                self.memory.write_bytes(FRAMEBUFFER_BASE, &clear)?;
-                self.memory
-                    .write_bytes(LEGACY_FRAMEBUFFER_ADDRESS, &clear)?;
-                cpu.r[0] = 0;
-            }
-            "FlushDCache" | "__dcache_writeback_all" => {
-                if self.memory.read_bytes(cpu.r[0], FRAMEBUFFER_SIZE).is_ok() {
-                    *self.active_framebuffer = cpu.r[0];
-                    *self.frame_address = Some(cpu.r[0]);
-                }
-                cpu.r[0] = 0;
-            }
-            "BMF_SelectPixelFunc" => cpu.r[0] = 0,
-            "LCDEnableDoubleFB" | "LCDDisableDoubleFB" | "LCDSetFBFormat" | "LCDInit"
-            | "LCDSetRefreshRate" | "LCDSetBrightness" | "InvalidICache" | "fsys_RefreshCache"
-            | "consoleEnable" | "consoleDisable" | "PMSetMode" => cpu.r[0] = 0,
-            _ => self.record_unknown(cpu, &symbol_name, symbol_address)?,
-        }
-        if self.profile == ArmProfile::Homebrew {
-            cpu.r[15] = cpu.r[14] & !1;
-        }
-        Ok(())
-    }
-
-    fn dispatch_semihosting(&mut self, cpu: &mut Cpu) -> Result<()> {
-        match cpu.r[0] {
-            0x03 => {
-                let value = self.memory.read8(cpu.r[1])?;
-                log::trace!("ARM semihosting SYS_WRITEC: {:?}", char::from(value));
-                self.append_console_output(&[value]);
-                cpu.r[0] = 0;
-            }
-            0x04 => {
-                let value = self.read_c_string(cpu.r[1], 4096)?;
-                log::trace!("ARM semihosting SYS_WRITE0: {value:?}");
-                self.append_console_output(value.as_bytes());
-                cpu.r[0] = 0;
-            }
-            0x18 | 0x20 => {
-                log::trace!(
-                    "ARM semihosting exit operation={:#04x}, reason={:#010x}",
-                    cpu.r[0],
-                    cpu.r[1]
-                );
-                self.stop_requested = true;
-                cpu.r[0] = 0;
-            }
-            operation => {
-                return Err(SimulatorError::SdkHleError(format!(
-                    "unsupported ARM semihosting operation {operation:#010x} with parameter {:#010x}",
-                    cpu.r[1]
-                )));
-            }
-        }
-        Ok(())
-    }
-
-    fn append_console_output(&mut self, output: &[u8]) {
-        if output.len() >= CONSOLE_OUTPUT_LIMIT {
-            self.console_output.clear();
-            self.console_output
-                .extend_from_slice(&output[output.len() - CONSOLE_OUTPUT_LIMIT..]);
-            return;
-        }
-        let overflow = self
-            .console_output
-            .len()
-            .saturating_add(output.len())
-            .saturating_sub(CONSOLE_OUTPUT_LIMIT);
-        if overflow != 0 {
-            self.console_output.drain(..overflow);
-        }
-        self.console_output.extend_from_slice(output);
-    }
-
-    fn allocate(&mut self, size: u32) -> u32 {
-        self.heap.allocate(size)
-    }
-
-    fn allocate_zeroed(&mut self, size: u32) -> Result<u32> {
-        let address = self.allocate(size);
-        if address != 0 && size != 0 {
-            self.memory.write_bytes(address, &vec![0; size as usize])?;
-        }
-        Ok(address)
-    }
-
-    fn deallocate(&mut self, address: u32) {
-        self.heap.deallocate(address);
-    }
-
-    fn reallocate(&mut self, address: u32, size: u32) -> Result<u32> {
-        self.heap.reallocate(self.memory, address, size)
-    }
-
-    fn dynamic_import(&mut self, name_address: u32) -> Result<u32> {
-        let name = self.read_c_string(name_address, 256)?;
-        let index = match self.dynamic_imports.iter().position(|item| item == &name) {
-            Some(index) => index,
-            None => {
-                self.dynamic_imports.push(name);
-                self.dynamic_imports.len() - 1
-            }
-        };
-        let address = DYNAMIC_THUNK_BASE + index as u32 * 8;
-        self.memory.write32(address, 0xef80_0000 | index as u32)?;
-        self.memory.write32(address + 4, 0xe12f_ff1e)?;
-        Ok(address)
-    }
-
-    fn open_file(&mut self, name: &str, mode: &str) -> u32 {
-        let operation = mode.as_bytes().first().copied().unwrap_or(b'r');
-        let writable = matches!(operation, b'w' | b'a') || mode.contains('+');
-        if !matches!(operation, b'r' | b'w' | b'a') {
-            log::trace!("ARM file open rejected mode: {name:?} ({mode:?})");
-            return 0;
-        }
-        let save_path = self
-            .save_directory
-            .and_then(|directory| resolve_guest_path(directory, name));
-        if writable && save_path.is_none() {
-            log::trace!("ARM file open has no safe save path: {name:?} ({mode:?})");
-            return 0;
-        }
-        if operation == b'w' {
-            log::trace!("ARM file create save path: {name:?}");
-            return self.insert_file(Vec::new(), save_path, 0, true, true);
-        }
-        if operation == b'a' {
-            let data = save_path
-                .as_ref()
-                .and_then(|path| std::fs::read(path).ok())
-                .unwrap_or_default();
-            let position = data.len();
-            log::trace!("ARM file append save path: {name:?}");
-            return self.insert_file(data, save_path, position, true, true);
-        }
-        if let Some((path, data)) = save_path
-            .as_ref()
-            .and_then(|path| std::fs::read(path).ok().map(|data| (path.clone(), data)))
-        {
-            log::trace!("ARM file open save path: {name:?} -> {}", path.display());
-            return self.insert_file(data, writable.then_some(path), 0, writable, false);
-        }
-        let Some(path) = resolve_guest_path(self.content_directory, name) else {
-            log::trace!("ARM file open rejected path: {name:?}");
-            return 0;
-        };
-        let data = match std::fs::read(&path) {
-            Ok(data) => {
-                log::trace!("ARM file open host path: {name:?} -> {}", path.display());
-                data
-            }
-            Err(error) => {
-                if let Some(resource) = self.package.find_resource(name) {
-                    log::trace!("ARM file open package resource: {name:?}");
-                    self.package.get_resource_data(resource)
-                } else if let Some(data) = self.package.get_embedded_file_data(name) {
-                    log::trace!("ARM file open appended package payload: {name:?}");
-                    data
-                } else if let Some(data) =
-                    self.firmware_archive.and_then(|archive| archive.read(name))
-                {
-                    log::trace!("ARM file open firmware resource: {name:?}");
-                    data
-                } else {
-                    log::trace!("ARM file open failed: {name:?} ({error})");
-                    return 0;
-                }
-            }
-        };
-        let persisted_path = if writable { save_path } else { None };
-        self.insert_file(data, persisted_path, 0, writable, false)
-    }
-
-    fn insert_file(
-        &mut self,
-        data: Vec<u8>,
-        save_path: Option<PathBuf>,
-        position: usize,
-        writable: bool,
-        dirty: bool,
-    ) -> u32 {
-        let stream = *self.next_file_handle;
-        *self.next_file_handle = stream.wrapping_add(1).max(1);
-        let handle = if self.profile == ArmProfile::Homebrew {
-            let address = self.allocate(16);
-            if address == 0 {
-                return 0;
-            }
-            if self.memory.write32(address, stream).is_err() {
-                self.deallocate(address);
-                return 0;
-            }
-            address
-        } else {
-            stream
-        };
-        self.files.insert(
-            handle,
-            GuestFile {
-                data,
-                position,
-                data_address: 0,
-                save_path,
-                writable,
-                dirty,
-            },
-        );
-        handle
-    }
-
-    fn open_resource(&mut self, candidates: [u32; 3]) -> u32 {
-        let found = candidates.into_iter().find_map(|address| {
-            if address < 0x1_0000 {
-                return None;
-            }
-            let name = self.read_c_string(address, 1024).ok()?;
-            let resource = self.package.find_resource(&name)?;
-            Some((name, self.package.get_resource_data(resource)))
-        });
-        let Some((name, data)) = found else {
-            return 0;
-        };
-        let handle = self.allocate(16);
-        if handle == 0 {
-            return 0;
-        }
-        log::trace!("ARM resource open {name:?} -> {handle:#010x}");
-        self.files.insert(
-            handle,
-            GuestFile {
-                data,
-                position: 0,
-                data_address: 0,
-                save_path: None,
-                writable: false,
-                dirty: false,
-            },
-        );
-        handle
-    }
-
-    fn read_resource(
-        &mut self,
-        handle: u32,
-        destination: u32,
-        buffer_len: u32,
-        read_len: u32,
-    ) -> Result<u32> {
-        if destination == 0 {
-            let existing = self.files.get(&handle).map_or(0, |file| file.data_address);
-            if existing != 0 {
-                return Ok(existing);
-            }
-            let Some(size) = self.files.get(&handle).map(|file| file.data.len() as u32) else {
-                return Ok(0);
-            };
-            let address = self.allocate(size);
-            if address == 0 {
-                return Ok(0);
-            }
-            let data = self.files[&handle].data.clone();
-            self.memory.write_bytes(address, &data)?;
-            self.files.get_mut(&handle).unwrap().data_address = address;
-            return Ok(address);
-        }
-
-        let data = {
-            let Some(file) = self.files.get_mut(&handle) else {
-                return Ok(0);
-            };
-            let available = file.data.len().saturating_sub(file.position);
-            let requested = if read_len != 0 && buffer_len > 1 {
-                (read_len as usize).saturating_mul(buffer_len as usize)
-            } else if read_len != 0 {
-                read_len as usize
-            } else {
-                buffer_len as usize
-            };
-            let length = if requested == 0 || requested > available {
-                available
-            } else {
-                requested
-            };
-            let data = file.data[file.position..file.position + length].to_vec();
-            file.position += length;
-            data
-        };
-        self.memory.write_bytes(destination, &data)?;
-        Ok(if read_len != 0 {
-            (data.len() / read_len as usize) as u32
-        } else {
-            data.len() as u32
-        })
-    }
-
-    fn read_file(&mut self, destination: u32, size: u32, count: u32, handle: u32) -> Result<u32> {
-        let Some(requested) = (size as usize).checked_mul(count as usize) else {
-            return Ok(0);
-        };
-        if size == 0 || requested == 0 {
-            return Ok(0);
-        }
-        let data = {
-            let Some(file) = self.files.get_mut(&handle) else {
-                return Ok(0);
-            };
-            if file.position >= file.data.len() {
-                return Ok(0);
-            }
-            let available = file.data.len().saturating_sub(file.position);
-            let length = requested.min(available);
-            let data = file.data[file.position..file.position + length].to_vec();
-            file.position += length;
-            data
-        };
-        self.memory.write_bytes(destination, &data)?;
-        Ok(data.len() as u32 / size)
-    }
-
-    fn write_file(&mut self, source: u32, size: u32, count: u32, handle: u32) -> Result<u32> {
-        let Some(requested) = (size as usize).checked_mul(count as usize) else {
-            return Ok(0);
-        };
-        if size == 0 || requested == 0 {
-            return Ok(0);
-        }
-        let data = self.memory.read_bytes(source, requested)?.to_vec();
-        let Some(file) = self.files.get_mut(&handle) else {
-            return Ok(0);
-        };
-        if !file.writable {
-            return Ok(0);
-        }
-        let Some(end) = file.position.checked_add(data.len()) else {
-            return Ok(0);
-        };
-        if file.data.len() < end {
-            file.data.resize(end, 0);
-        }
-        file.data[file.position..end].copy_from_slice(&data);
-        file.position = end;
-        file.dirty = true;
-        Ok(count)
-    }
-
-    fn close_file(&mut self, handle: u32) -> u32 {
-        let Some(file) = self.files.get_mut(&handle) else {
-            return u32::MAX;
-        };
-        let result = flush_guest_file(file);
-        self.files.remove(&handle);
-        self.deallocate(handle);
-        if let Err(error) = result {
-            log::error!("Failed to close ARM guest save file {handle}: {error}");
-            u32::MAX
-        } else {
-            0
-        }
-    }
-
-    fn begin_file_search(
-        &mut self,
-        pattern: &str,
-        attributes: u32,
-        data_address: u32,
-    ) -> Result<u32> {
-        self.file_searches.remove(&data_address);
-        if data_address == 0 {
-            return Ok(u32::MAX);
-        }
-        let Some(entries) = self.collect_file_search_entries(pattern, attributes) else {
-            return Ok(u32::MAX);
-        };
-        let Some(first) = entries.first().cloned() else {
-            return Ok(u32::MAX);
-        };
-        self.write_file_search_name(data_address, &first)?;
-        self.file_searches.insert(
-            data_address,
-            FileSearch {
-                entries,
-                next_index: 1,
-            },
-        );
-        Ok(0)
-    }
-
-    fn next_file_search(&mut self, data_address: u32) -> Result<u32> {
-        let Some(name) = self
-            .file_searches
-            .get_mut(&data_address)
-            .and_then(|search| {
-                let name = search.entries.get(search.next_index)?.clone();
-                search.next_index += 1;
-                Some(name)
-            })
-        else {
-            return Ok(u32::MAX);
-        };
-        self.write_file_search_name(data_address, &name)?;
-        Ok(0)
-    }
-
-    fn collect_file_search_entries(&self, pattern: &str, attributes: u32) -> Option<Vec<String>> {
-        let (directory, file_pattern) = normalize_guest_search_pattern(pattern)?;
-        let root = self.content_directory.canonicalize().ok()?;
-        let search_directory = if directory.as_os_str().is_empty() {
-            root.clone()
-        } else {
-            root.join(directory).canonicalize().ok()?
-        };
-        if !search_directory.starts_with(&root) {
-            return None;
-        }
-
-        let include_directories = attributes & 0x10 != 0;
-        let include_files = attributes & 0x20 != 0 || !include_directories;
-        let mut entries = std::fs::read_dir(search_directory)
-            .ok()?
-            .filter_map(|entry| entry.ok())
-            .filter_map(|entry| {
-                let file_type = entry.file_type().ok()?;
-                if (file_type.is_dir() && !include_directories)
-                    || (file_type.is_file() && !include_files)
-                    || (!file_type.is_dir() && !file_type.is_file())
-                {
-                    return None;
-                }
-                let name = entry.file_name().to_string_lossy().into_owned();
-                wildcard_matches(&file_pattern, &name).then_some(name)
-            })
-            .collect::<Vec<_>>();
-        entries.sort_by(|left, right| {
-            left.to_ascii_lowercase()
-                .cmp(&right.to_ascii_lowercase())
-                .then_with(|| left.cmp(right))
-        });
-        Some(entries)
-    }
-
-    fn write_file_search_name(&mut self, data_address: u32, name: &str) -> Result<()> {
-        let name = name.as_bytes();
-        let length = name.len().min(FILE_SEARCH_NAME_CAPACITY - 1);
-        let destination = data_address.wrapping_add(FILE_SEARCH_NAME_OFFSET);
-        self.memory.write_bytes(destination, &name[..length])?;
-        self.memory.write8(destination + length as u32, 0)
-    }
-
-    fn seek_file(&mut self, handle: u32, offset: i32, origin: u32) -> u32 {
-        let Some(file) = self.files.get_mut(&handle) else {
-            return u32::MAX;
-        };
-        let base = match origin {
-            0 => 0,
-            1 => file.position as i128,
-            2 => file.data.len() as i128,
-            _ => return u32::MAX,
-        };
-        let position = base + i128::from(offset);
-        let Ok(position) = usize::try_from(position) else {
-            return u32::MAX;
-        };
-        file.position = position;
-        0
-    }
-
-    fn read_c_string(&self, address: u32, limit: usize) -> Result<String> {
-        let mut bytes = Vec::new();
-        for offset in 0..limit {
-            let value = self.memory.read8(address.wrapping_add(offset as u32))?;
-            if value == 0 {
-                return Ok(String::from_utf8_lossy(&bytes).into_owned());
-            }
-            bytes.push(value);
-        }
-        Err(SimulatorError::SdkHleError(
-            "unterminated ARM guest string".into(),
-        ))
-    }
-
-    fn read_wide_string(&self, address: u32, limit: usize) -> Result<String> {
-        let mut words = Vec::new();
-        for offset in 0..limit {
-            let value = self
-                .memory
-                .read16(address.wrapping_add((offset * 2) as u32))?;
-            if value == 0 {
-                return Ok(String::from_utf16_lossy(&words));
-            }
-            words.push(value);
-        }
-        Err(SimulatorError::SdkHleError(
-            "unterminated ARM guest wide string".into(),
-        ))
-    }
-
-    fn write_guest_string(&mut self, address: u32, capacity: u32, value: &str) -> Result<bool> {
-        if address == 0 || capacity == 0 {
-            return Ok(false);
-        }
-        let count = value.len().min(capacity.saturating_sub(1) as usize);
-        self.memory
-            .write_bytes(address, &value.as_bytes()[..count])?;
-        self.memory.write8(address + count as u32, 0)?;
-        Ok(count == value.len())
-    }
-
-    fn record_unknown(&mut self, cpu: &mut Cpu, name: &str, import_address: u32) -> Result<()> {
-        let pc = cpu.r[15].wrapping_sub(4);
-        let call = self
-            .unknown_hle_calls
-            .entry(name.to_string())
-            .or_insert_with(|| UnknownHleCall {
-                name: name.to_string(),
-                count: 0,
-                import_address,
-                first_pc: pc,
-                first_arguments: [cpu.r[0], cpu.r[1], cpu.r[2], cpu.r[3]],
-            });
-        call.count += 1;
-        if self.unknown_hle_policy == UnknownHlePolicy::Stop
-            && !self.unknown_hle_allowlist.contains(name)
-        {
-            return Err(SimulatorError::UnknownHle {
-                name: name.to_string(),
-                pc,
-                import_address,
-                arguments: [cpu.r[0], cpu.r[1], cpu.r[2], cpu.r[3]],
-            });
-        }
-        cpu.r[0] = 0;
-        Ok(())
-    }
-}
-
 fn framebuffer_is_solid(framebuffer: &[u8]) -> bool {
     let pixels = framebuffer.as_chunks::<2>().0;
     pixels
@@ -1976,46 +1131,6 @@ fn compare_ascii_case_insensitive(left: &str, right: &str) -> i32 {
         std::cmp::Ordering::Less => -1,
         std::cmp::Ordering::Equal => 0,
         std::cmp::Ordering::Greater => 1,
-    }
-}
-
-impl Bus for RuntimeBus<'_> {
-    fn read8(&mut self, address: u32) -> Result<u8> {
-        self.memory.read8(address)
-    }
-    fn read16(&mut self, address: u32) -> Result<u16> {
-        self.memory.read16(address)
-    }
-    fn read32(&mut self, address: u32) -> Result<u32> {
-        self.memory.read32(address)
-    }
-    fn write8(&mut self, address: u32, value: u8) -> Result<()> {
-        self.memory.write8(address, value)
-    }
-    fn write16(&mut self, address: u32, value: u16) -> Result<()> {
-        self.memory.write16(address, value)
-    }
-    fn write32(&mut self, address: u32, value: u32) -> Result<()> {
-        self.memory.write32(address, value)?;
-        if address == LEGACY_GRAPHICS_STRIDE {
-            match value {
-                value if value == SCREEN_WIDTH * 2 => {
-                    *self.framebuffer_bits = 16 | FRAMEBUFFER_BITS_EXPLICIT;
-                }
-                value if value == SCREEN_WIDTH * 4 => {
-                    *self.framebuffer_bits = 32 | FRAMEBUFFER_BITS_EXPLICIT;
-                }
-                _ => {}
-            }
-        }
-        if address == LEGACY_GRAPHICS_SURFACE {
-            *self.active_framebuffer = LEGACY_FRAMEBUFFER_ADDRESS;
-            *self.frame_address = Some(LEGACY_FRAMEBUFFER_ADDRESS);
-        }
-        Ok(())
-    }
-    fn svc(&mut self, cpu: &mut Cpu, immediate: u32) -> Result<()> {
-        self.dispatch(cpu, immediate)
     }
 }
